Specific thrust calibration state
Introduces a hover pause before the thrust calibration.
With that it can be avoided that the thrust calibration is executed with the drone still hovering stable at the previous waypoint.
The new hover pause give the drone some time to react on missing y-integrator and such it becomes clearer if the thrust needs to be adjusted or not.