[landing] Remove opportunity of hunting in landing
Prevents bad landings where the drone just tries to hunt shortly before
touch down. Drone still looses height and tracking gets lost.
At the end the drone seems to land, with a bad yaw but it is not
starting to shake because dronenavigation is in the wrong state
(ns_chasing).