Switch targets based on TTI
With the rapid route optimizer we can run the optimization on all insects simultaneously. This behaves similarly to selecting the target based on acceleration required, but considers the estimated interception position rather than the extrapolated insect position over an arbitrary time AND considers whether or not both the interception position and the stopping position are within 'view'.