Avoid optimizer results at full deceleration thrust
If the drone has achieved significant velocity, a second solution to the binary search is to decelerate at full throttle. This commit tries to avoid finding this solution by comparing the direction of the acceleration vector and position error to the interception point. If these point in opposite directions, reduce the allowed time for interception to hopefully flip the vector.