Consider a point exactly on a plane to be inside it. Initialize ref_distance as max possible value.
Aligns with the "move inside" functions' behavior, where in certain cases the point is placed right on the volume boundary, currently failing some unit tests.

By initializing the ref_distance as max we can support moving points from as far away as the numeric limits of a float, rather than "just" a distance of 999.