[x] Remove dedicated take-off score threshold.
Although this can prevent tracking the wrong thing at take-off it was much too strict causing many failed take-off detects.

If in the future we want to treat the take-off as a special case again, it should function similar to the regular score threshold, i.e. grow as long as we do not track the drone.